+++ src/frontend/packages/lib_components/src/types/index.ts
@@ -10,7 +10,6 @@ export * from './PortabilityRequest';
 export * from './Organization';
 export * from './ResourceContext';
 export * from './RouteOptions';
-export * from './serviceWorker';
 export * from './tracks';
 export * from './User';
 export * from './window';