+++ src/frontend/packages/lib_components/src/utils/serviceWorker/index.ts
@@ -1,3 +1,1 @@
 export * as serviceWorkerRegistration from './serviceWorkerRegistration';
-export * as serviceWorkerForceInstall from './serviceWorkerForceInstall';
-export * from './serviceWorkerRefreshToken';